Add minimal support for dns-01
Add support for specifying challenge type, and add hook script support
for type dns-01.  The hook syntax is compatibel with letsencrypt.sh

+++ acme_tiny.py
@@ -12,7 +12,7 @@ LOGGER = logging.getLogger(__name__)
 LOGGER.addHandler(logging.StreamHandler())
 LOGGER.setLevel(logging.INFO)
 
-def get_crt(account_key, csr, acme_dir, log=LOGGER, CA=DEFAULT_CA):
+def get_crt(account_key, csr, acme_dir, log=LOGGER, CA=DEFAULT_CA, challenge_type='http-01', hook=None):
     # helper function base64 encode for jose spec
     def _b64(b):
         return base64.urlsafe_b64encode(b).decode('utf8').replace("=", "")
@@ -104,23 +104,32 @@ def get_crt(account_key, csr, acme_dir, log=LOGGER, CA=DEFAULT_CA):
             raise ValueError("Error requesting challenges: {0} {1}".format(code, result))
 
         # make the challenge file
-        challenge = [c for c in json.loads(result.decode('utf8'))['challenges'] if c['type'] == "http-01"][0]
+        print result.decode('utf8')
+        challenge = [c for c in json.loads(result.decode('utf8'))['challenges'] if c['type'] == challenge_type][0]
         token = re.sub(r"[^A-Za-z0-9_\-]", "_", challenge['token'])
         keyauthorization = "{0}.{1}".format(token, thumbprint)
-        wellknown_path = os.path.join(acme_dir, token)
-        with open(wellknown_path, "w") as wellknown_file:
-            wellknown_file.write(keyauthorization)
 
-        # check that the file is in place
-        wellknown_url = "http://{0}/.well-known/acme-challenge/{1}".format(domain, token)
-        try:
-            resp = urlopen(wellknown_url)
-            resp_data = resp.read().decode('utf8').strip()
-            assert resp_data == keyauthorization
-        except (IOError, AssertionError):
-            os.remove(wellknown_path)
-            raise ValueError("Wrote file to {0}, but couldn't download {1}".format(
-                wellknown_path, wellknown_url))
+        if challenge_type == 'http-01':
+            wellknown_path = os.path.join(acme_dir, token)
+            with open(wellknown_path, "w") as wellknown_file:
+                wellknown_file.write(keyauthorization)
+            # check that the file is in place
+            wellknown_url = "http://{0}/.well-known/acme-challenge/{1}".format(domain, token)
+            try:
+                resp = urlopen(wellknown_url)
+                resp_data = resp.read().decode('utf8').strip()
+                assert resp_data == keyauthorization
+            except (IOError, AssertionError):
+                os.remove(wellknown_path)
+                raise ValueError("Wrote file to {0}, but couldn't download {1}".format(
+                    wellknown_path, wellknown_url))
+        elif challenge_type == 'dns-01':
+            if hook is None:
+                raise ValueError("Missing hook script")
+            print keyauthorization
+            subprocess.check_call([hook, "deploy_challenge", domain, token, _b64(hashlib.sha256(keyauthorization.encode('utf8')).digest())]),
+        else:
+            raise ValueError("Unknown challenge type {0}".format(challenge_type))
 
         # notify challenge are met
         code, result = _send_signed_request(challenge['uri'], {
@@ -142,7 +151,11 @@ def get_crt(account_key, csr, acme_dir, log=LOGGER, CA=DEFAULT_CA):
                 time.sleep(2)
             elif challenge_status['status'] == "valid":
                 log.info("{0} verified!".format(domain))
-                os.remove(wellknown_path)
+                if challenge_type == 'http-01':
+                    os.remove(wellknown_path)
+                elif challenge_type == 'dns-01':
+                    subprocess.check_call([hook, "deploy_challenge", domain, token, _b64(hashlib.sha256(keyauthorization.encode('utf8')).digest())]),
+
                 break
             else:
                 raise ValueError("{0} challenge did not pass: {1}".format(
@@ -188,10 +201,11 @@ def main(argv):
     parser.add_argument("--acme-dir", required=True, help="path to the .well-known/acme-challenge/ directory")
     parser.add_argument("--quiet", action="store_const", const=logging.ERROR, help="suppress output except for errors")
     parser.add_argument("--ca", default=DEFAULT_CA, help="certificate authority, default is Let's Encrypt")
-
+    parser.add_argument("--challenge-type", default="http-01", help="What challenge type to use")
+    parser.add_argument("--hook", default=None, help="What hook to use (only used for DNS challenges)")
     args = parser.parse_args(argv)
     LOGGER.setLevel(args.quiet or LOGGER.level)
-    signed_crt = get_crt(args.account_key, args.csr, args.acme_dir, log=LOGGER, CA=args.ca)
+    signed_crt = get_crt(args.account_key, args.csr, args.acme_dir, log=LOGGER, CA=args.ca, challenge_type=args.challenge_type, hook=args.hook)
     sys.stdout.write(signed_crt)
 
 if __name__ == "__main__": # pragma: no cover
